pyproject.toml
[build-system]
requires = ["uv_build>=0.8.22"]
build-backend = "uv_build"

[project]
name = "ucl-open-vr-corridor-2p"
description = "A library that defines a ucl-open data schema for the vr-corridor-2p experiment."
authors = [ {name = "RoboDoig", email = "a.erskine@neurogears.org"}]
license = "MIT"
requires-python = ">=3.11, <3.12"
classifiers = [
    "Programming Language :: Python :: 3.11",
    "Operating System :: Microsoft :: Windows",
]
version = "0.0.0-rc1"
readme = {file = "README.md", content-type = "text/markdown"}

dependencies = [
  "swc-aeon>=0.1.0",
   "ucl-open-rigs",
  "pydantic-settings"
]

# [project.urls]
# Documentation = "https://ucl-open.github.io/Ucl.Open.VrCorridor2p/"
# Repository = "https://github.com/UclOpen/Ucl.Open.VrCorridor2p/"
# Issues = "https://github.com/UclOpen/Ucl.Open.VrCorridor2p/issues"
# Changelog = "https://github.com/UclOpen/Ucl.Open.VrCorridor2p/releases"
[project.scripts]
vr-corridor-2p = "ucl_open_vr_corridor_2p.cli:main"

[tool.ruff]
line-length = 120
target-version = 'py311'

[tool.ruff.lint]
extend-select = ['Q', 'RUF100', 'C90', 'I']
extend-ignore = []
mccabe = { max-complexity = 14 }
pydocstyle = { convention = 'google' }

[tool.codespell]
skip = '.git,*.pdf,*.svg,./bonsai,*.bonsai,./docs/_build,uv.lock,*Generated.cs'
ignore-words-list = 'nd'

[tool.uv]
required-version = '>=0.8.4'

[tool.uv.sources]
ucl-open-rigs = {git = "https://github.com/ucl-open/ucl_open_rigs/", branch = "test_build"}
swc-aeon = { git = "https://github.com/SainsburyWellcomeCentre/aeon_api.git", branch = "pydantic-schema" }